Fix InstrumentDefMsg Char byte alignment and MBP10Msg tuple creation
1. InstrumentDefMsg Char fields fix (src/decode.jl, src/encode.jl):
   - Changed match_algorithm and security_update_action to read/write as UInt8
   - Julia's read(io, Char) can read 1-4 bytes for UTF-8, causing misalignment
   - DBN format uses single-byte C char (always 1 byte)
   - Now: read UInt8, convert to Char only if non-zero
   - Encoder updated to write UInt8(char) instead of char directly
   - This fixes "Read more bytes than expected: 394 vs 384" (10 bytes = 2 Chars reading as multi-byte)

2. MBP10Msg tuple creation fix (test/compatibility_utils.jl):
   - Use ntuple with named helper function instead of closure
   - Avoids world age issues and closure capture problems
   - Helper function make_level() creates BidAskPair from JSON or zeros
   - Clearer and more reliable than closure: ntuple(i -> levels_vec[i], 10)

Both changes maintain compatibility with Julia's own serialization while
fixing compatibility with Rust-generated files.

+++ src/encode.jl
@@ -403,12 +403,12 @@ function write_record(encoder::DBNEncoder, record)
         
         write(io, UInt8(record.instrument_class))
         write(io, record.strike_price)
-        write(io, record.match_algorithm)
+        write(io, UInt8(record.match_algorithm))  # Write as UInt8, not Char
         write(io, record.main_fraction)
         write(io, record.price_display_format)
         write(io, record.sub_fraction)
         write(io, record.underlying_product)
-        write(io, record.security_update_action)
+        write(io, UInt8(record.security_update_action))  # Write as UInt8, not Char
         write(io, record.maturity_month)
         write(io, record.maturity_day)
         write(io, record.maturity_week)
